+++ scripts/build-linux-gcc.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+libedhoc_options="-DLIBEDHOC_ENABLE_UNIT_TESTS=ON -DLIBEDHOC_BUILD_COMPILER_GCC=ON"
+mbedtls_options="-DENABLE_PROGRAMS=OFF -DENABLE_TESTING=OFF"
+
+rm -rf build
+mkdir build
+cd build
+cmake $libedhoc_options $mbedtls_options ..
+make -j