+++ scripts/zcbor_conv_cddls_to_c.py
@@ -1,6 +1,15 @@
 import subprocess
 import os
 
+"""
+## Steps to run script zcbor_conv_cddls_to_c.sh:
+- `python3 -m venv env`
+- `source env/bin/activate`
+- `pip3 install -r ../externals/zcbor/scripts/requirements-base.txt`
+- `python3 zcbor_conv_cddls_to_c.py`
+- `deactivate`
+"""
+
 # Define paths and models
 zcbor_py = "../externals/zcbor/zcbor/zcbor.py"
 models = {